+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 - Changed TABLE_MAP_EVENT will no longer appear after adding events to only/ignore configuration 
 - Fixed events with dropped columns will return a proper columns amount
 - Changed configuration to static calls
+- Removed absolute method getConnection from repository
 
 ## v2.2.0 (2017-03-10)
 - Removed foreign keys from events